Login required and user registeration.

+++ python/django/studybud/base/views.py
@@ -1,16 +1,22 @@
-from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, login, logout
-from django.db.models import Q
-from django.shortcuts import redirect, render
 from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.db.models import Q
+from django.http import HttpResponse
+from django.shortcuts import redirect, render
+from django.contrib.auth.forms import UserCreationForm
 
 from .forms import RoomForm
 from .models import Room, Topic
 
 
 def login_page(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
-        username = request.POST.get("username")
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
 
         try:
@@ -27,13 +33,33 @@ def login_page(request):
                 return redirect("home")
             else:
                 messages.error(request, "username or password is incorrect.")
-    context = {}
+    context = {"page": "login"}
     return render(request, "base/login_register.html", context)
 
 def logout_user(request):
     logout(request)
     return redirect("home")
 
+def register_page(request):
+    form = UserCreationForm()
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error occured during registeration.")
+
+        
+
+            
+
+
+    context = {"form": form}
+    return render(request, "base/login_register.html", context)
 
 def home(request):
     q = request.GET.get('q')
@@ -55,6 +81,7 @@ def room(request, pk):
     context = {"room": room}
     return render(request, "base/room.html", context)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -67,9 +94,12 @@ def create_room(request):
     context = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("<h1>You are not allowed here!</h1>")
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -78,8 +108,11 @@ def update_room(request, pk):
     context = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("<h1>You are not allowed here!</h1>")
     if request.method == "POST":
         room.delete()
         return redirect("home")
